fix(davinci): keep assurance harness files under the source-length gate
